began refactoring the main page
