+++ app/components/SphinxContainer.jsx
@@ -1,29 +1,32 @@
 
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
-import Cat from './Cat';
-import ChatBox from './ChatBox';
+import Sphinx from './Sphinx';
+import Riddle from './Riddle';
 import Inf from './Inf';
 import {addMessage} from '../reducers/chat';
 
 
-class CatContainer extends Component {
+class SphinxContainer extends Component {
 
   constructor(props) {
     super(props)
   }
   render() {
-    console.log(this.props)
     return(
       <div className="home">
         <div className="chat">
-          <ChatBox
-            messages={this.props.messages}
-            addMessage={this.props.dispatchAddMessage}
+          <Riddle
+            //hard-code props until we set up the database
+            currentRiddle={"What is the creature that walks on four legs in the morning, two legs at noon and three in the evening?"}
+
+            solution={["human", "man", "person"]}
+            userTried={true}
+            userGuessedCorrectly={true}
           />
         </div>
         <div className="cat">
-          <Cat/>
+          <Sphinx />
         </div>
       </div>
     )
@@ -47,5 +50,5 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(CatContainer);
+export default connect(mapStateToProps, mapDispatchToProps)(SphinxContainer);
