+++ app/main.jsx
@@ -12,7 +12,7 @@ import Welcome from './components/Welcome';
 import Signup from './components/Signup';
 import Login from './components/Login';
 
-import CatContainer from './components/CatContainer';
+import SphinxContainer from './components/SphinxContainer';
 import ImageAPI from './components/ImageAPI';
 import Chat from './components/Chat';
 
@@ -23,6 +23,12 @@ import injectTapEventPlugin from 'react-tap-event-plugin';
 // http://stackoverflow.com/a/34015469/988941
 injectTapEventPlugin();
 
+// After we implement the database, we will need to write a reducer "getCurrentRiddle" that retrieves a riddle from the database and places it in the store.
+
+// const onSphinxContainerEnter = function (nextRouterState) {
+//   store.dispatch(getCurrentRiddle);
+// };
+
 render(
   <MuiThemeProvider>
     <Provider store={store}>
@@ -32,7 +38,7 @@ render(
           <Route path="/home" component={Welcome} />
           <Route path="/signup" component={Signup} />
           <Route path="/login" component={Login} />
-          <Route path="/cat" component={CatContainer} />
+          <Route path="/sphinx" component={SphinxContainer} />
           <Route path="/images" component={ImageAPI} />
           <Route path="/chat" component={Chat} />
         </Route>
