+++ public/index.html
@@ -2,7 +2,7 @@
 
 <html lang="en">
   <head>
-    <title>Virtual Cat</title>
+    <title>Sphinx</title>
     <link href='/normalize.css' rel='stylesheet' >
     <link href="https://fonts.googleapis.com/css?family=Roboto" rel="stylesheet">
     <link href="/style.css" rel="stylesheet" />
